Handle variable framerates
The previous implementation assumed framerates were consistent,
which isn't the case if we drop frames due to RTSP dropout or similar.

When this happens, we don't get a framerate fps value from the
video reader.

To calculate the output FPS for the written video, calculate by
reading every video frame and corresponding timestamp.

Then for frame annotation, use the caps value returned for the
time offset of each frame.

One other minor change in this commit to get rid of warnings for
invalid codec when using caps version of MP4V string instead of
lowercase.  This likely wasn't related but is probably the right
change to make