Handle multiple events per frame
When frames are dropped in the video this looks like multiple
events which occur on a shorter range of frames.

When this happens, place the annotation on whatever frame happens
before the event in the case that the next frame is after the event.

A better way might be to try to find the closest frame in
timestamp however the approach proposed here is simpler and
works when the total time gap of dropped frames is relatively small.